requirements.txt
selenium
pytest
selenium-authenticated-proxy
pyautogui
requests
allure-pytest